move PPL_FIELD_TYPES to data plugin
