+++ src/plugins/explore/opensearch_dashboards.json
@@ -18,8 +18,7 @@
     "usageCollection",
     "visualizations",
     "expressions",
-    "dashboard",
-    "queryEnhancements"
+    "dashboard"
   ],
   "requiredBundles": ["opensearchDashboardsReact", "opensearchDashboardsUtils", "dashboard"],
   "optionalPlugins": ["home", "share"],
