+++ src/plugins/query_enhancements/common/types.ts
@@ -75,24 +75,3 @@ export enum S3_FIELD_TYPES {
   ARRAY = 'array',
   UNKNOWN = 'unknown', // For unmapped or unsupported types
 }
-
-// Types from opensearch-project/sql/docs/user/ppl/general/datatypes.rst
-export enum PPL_FIELD_TYPES {
-  BOOLEAN = 'boolean',
-  TINYINT = 'tinyint',
-  SMALLINT = 'smallint',
-  INT = 'int',
-  BIGINT = 'bigint',
-  FLOAT = 'float',
-  DOUBLE = 'double',
-  STRING = 'string',
-  TIMESTAMP = 'timestamp',
-  DATE = 'date',
-  TIME = 'time',
-  INTERVAL = 'interval',
-  IP = 'ip',
-  GEO_POINT = 'geo_point',
-  BINARY = 'binary',
-  STRUCT = 'struct',
-  ARRAY = 'array',
-}
