+++ src/plugins/query_enhancements/public/index.ts
@@ -12,4 +12,3 @@ export function plugin(initializerContext: PluginInitializerContext) {
 }
 
 export { QueryEnhancementsPluginSetup, QueryEnhancementsPluginStart } from './types';
-export { PPL_FIELD_TYPES, S3_FIELD_TYPES } from '../common/types';
